requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.3
Pillow==9.5.0
python-docx==0.8.11
requests==2.31.0
urllib3==2.0.3
Werkzeug==2.3.6
